% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get Dataset in krdata}
\usage{
get_data(dataId, dataList)
}
\arguments{
\item{dataId}{ID what you want to get from list.}

\item{dataList}{list of data from get_list function.}
}
\value{
Returns True if the data was successfully loaded.
}
\description{
Get dataset registered in krdata.
}
